Strix 1.6.0
Feats
1. 依赖升级。
2. tomcat -> undertow.
3. java 17 -> java 21.
4. gradle 8.3 -> gradle 8.4.
5. jdbc url 添加 `rewriteBatchedStatements=true`.
6. logback 开发环境不保留文件日志.
7. `KeysDiffHandler` 增加 `updatedFunc` 任意变动时回调.
8. 代码规范优化.

Fixes
empty.

+++ src/main/java/cn/projectan/strix/service/impl/OssFileServiceImpl.java
@@ -144,7 +144,7 @@ public class OssFileServiceImpl extends ServiceImpl<OssFileMapper, OssFile> impl
             byte[] imageByte = Base64.getDecoder().decode(data);
             for (int i = 0; i < imageByte.length; ++i) {
                 if (imageByte[i] < 0) {
-                    imageByte[i] += 256;
+                    imageByte[i] += (byte) 256;
                 }
             }
